refactor: OOP ftw
