+++ node/tsconfig.json
@@ -5,7 +5,6 @@
     "module": "commonjs",
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
-    "baseUrl": ".",
     "strict": true
   }
 }
